backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-google-genai
pypdf
python-dotenv
faiss-cpu